Create backend php
Install doctrine and connect our app to the sqlite database.
Create the entity with all the necessary fields (id, timestamp).
Create the controller to connect that entity with the
database depending on the request given by the user (GET or POST).
Make the migrations and test that the two methods work correctly.